[FIX] runbot: remove useless commit()
This commit() is the root cause of the famous false-positive error
`_check_module_names invalid module names, ignored: False`.

The first step (pull code) is a long one. if a push-hook request
happen during this step, the current transaction wont be able to be
commited (the list of module to test is written on the build). Due to
this early commit(), the build is left in an intermediate state with
the module list unpopulated (=False).