[FIX] runbot: kill builds when unlinking repo or branch
When a repo is deleted, the corresponding builds processes continues to
run indefinitely. Because of that, it may happens that the runbot launch
a new build using a port that is still in use by such a zombie build.

If this commit is applied, when unlinking a repo or a branch, the
corresponding builds are marked to be killed. Also, when searching for a
free port, the port is checked to avoid 'port already in use'
tracebacks.