[IMP] runbot: use cleanup to remove dirs
If the setUp method fails, the tearDown method is never called. This
prevent temporary dirs to be removed.

With this commit, cleanUp methods are added to be sure that the
temporary dirs are removed.

Also, as no builds are created during the test, we don't try to remove
dirs.

The repo attribute was defined two times, one in the setUp method, where
it was defined as the runbot.repo model and in the test itself where it
was defined as a runbot.repo instance. With this commit, it's defined
only once to avoid confusion.