[IMP] runbot: allow extra builds based on a cron
Actually, a build in a branch is created when a new commit is found.
When a user wants to schedule a time-based build, he has to generate a
commit outside of the runbot (e.g. in a cron script).

With this commit, cron builds can be scheduled daily with extra parameters.
By default, extra builds are started at 1 am UTC.

Also, this commit introduce the coverage result. If a cronbuild has the
coverage flag, the generated build is also flagged.
After the build job, a new job reads the '.coverage' files and store the
result on the build object.

This commit also adds a new computed field on the branch model that
retrieve the last coverage value from the builds.