[FIX] runbot: force reverse build even when testing
When a reverse build is needed in another repository, the _force method
is called on the latest build found. If the latest build is in testing
or pending state, nothing happens.

With this commit, the _force method will duplicate the build even in
testing 'state'. If the build is pending, the dependency build is
updated accordingly.

Also, the logs were wrongly referencing the latest builds as the forced
build, creating confusion.