[IMP] runbot: add rebuild button on build page
When examining a particular build with the build view, one can be
frustrated being forced to navigate the frontend page to ask for a
rebuild.

With this commit, the rebuild menu entry is also visible on the build
page when the build is the last one of the branch.

Also the build host is now visible on the build page for the same
usability reason.