[FIX] runbot: use regular branches to test last resort
To test the last resort branch matching when nothing in common can be
found, two PR were used leading to PR's target branch as the default
one.

Also, the test was never run beacause of a bad indentation.

With this commit, the indentation is fixed and the test uses regular
branches.