[FIX] runbot: avoid recursive duplicate repos
When defining a duplicate repository on repo A pointing to repo B, it
happens that repo B already has a duplicate repository pointing to A.
In that case, an infinite loop may happen.

With this commit, a validation error is raised whenever that happens.

Note: two other test are fixed here because they were using that kind of
recursive reference.