[FIX] runbot: write build results in a separate job
When a build reach the job30_run method, results from a previous testing
methods are computed.
With the previous commit 8c73e6a90133c8915ced619a31e48b10fb01b70d this
job can now be skipped. In that case, the results are not set.

With this commit, the results are computed in a separate method.