[REF] runbot: proof of concept standalone script
Actually, when deploying multiple instances of runbot, a full odoo
installation is needed.

With this commit, a single standalone (sort of) can be used to start a
runbot builder. It uses the same database as the runbot scheduler.
This script just reserve pending builds for himself and starts to build.

This is just a proof of concept. It should work alongside with a
standard runbot installation.